Changed data type long long to just long.

+++ balancedEvenSumRange.c
@@ -12,7 +12,7 @@
 typedef enum {RED,BLACK} color_t;
 
 typedef struct node_t {
-  long long value;
+  long value;
   color_t color;
   struct node_t *parent;
   struct node_t *left_child;
@@ -71,6 +71,7 @@ node_t *sibling(node_t *x) {
   }
 }
 
+//transplant
 void transplant(node_t **t, node_t *u, node_t *v) {
   node_t *p =u->parent;
 
@@ -89,6 +90,7 @@ void transplant(node_t **t, node_t *u, node_t *v) {
   }
 }
 
+//rotateLeft
 void rotateLeft(node_t **t,node_t *x) {
   node_t *y = x->right_child;
   node_t *b = y->left_child;
@@ -103,6 +105,7 @@ void rotateLeft(node_t **t,node_t *x) {
   x->parent = y;
 }
 
+//rotateRight
 void rotateRight(node_t **t, node_t *x) {
   node_t *y = x->left_child;
   node_t *b = y->right_child;
@@ -117,7 +120,7 @@ void rotateRight(node_t **t, node_t *x) {
   x->parent = y;
 }
 
-//rbInsertFixup(T,z)
+//rbInsertFixupA(T,z)
 void rbInsertFixupA(node_t **t, node_t *z) {
   node_t *y;
   while (z != *t && (z->parent)->color == RED) {
@@ -132,6 +135,7 @@ void rbInsertFixupA(node_t **t, node_t *z) {
   }
 }
 
+//rbInsertFixupC(T,z)
 void rbInsertFixupC(node_t **t, node_t *z) {
   node_t *x,*w;
   if (z == *t || (z->parent)->color == BLACK) {
@@ -153,6 +157,7 @@ void rbInsertFixupC(node_t **t, node_t *z) {
   }
 }
 
+//rbInsertFixupB(T,z)
 void rbInsertFixupB(node_t **t, node_t *z) {
   node_t *x,*w;
   if (z == *t || (z->parent)->color == BLACK) {
@@ -173,6 +178,7 @@ void rbInsertFixupB(node_t **t, node_t *z) {
   rbInsertFixupB(t,z);
 }
 
+//rbInsertFixup(T,z)
 void rbInsertFixup(node_t **t,node_t *z) {
   rbInsertFixupA(t,z);
   rbInsertFixupB(t,z);
@@ -181,7 +187,7 @@ void rbInsertFixup(node_t **t,node_t *z) {
 }
 
 // Inserts a new node into binary tree.
-void btreeInsert(node_t **t, long long n) {
+void btreeInsert(node_t **t, long n) {
   // Allocate memory for new node being inserted and initialize pointer for
   // the appropriate parent.
   node_t *z = malloc(sizeof(node_t));
@@ -214,10 +220,10 @@ void btreeInsert(node_t **t, long long n) {
 }
 
 // btreeEvenSumRange takes in the root node pointer of a binary tree and the
-// minimum and maximum of a range as long long.  It returns a 1 if the sum of
+// minimum and maximum of a range as long.  It returns a 1 if the sum of
 // the values within the range is odd, a 0 if it is even, and a -1 if there are
 // no values found within the given range.
-int btreeEvenSumRange(node_t *r, long long min, long long max) {
+int btreeEvenSumRange(node_t *r, long min, long max) {
   // Initialize i, left, and right variables to -1
   int i = -1;
   int left = -1;
@@ -282,26 +288,26 @@ int main(int argc,char *argv[]) {
   FILE *data;             // Initialize FILE pointers for data and ranges
   FILE *ranges;
 
-  long long inputData;    // Temporary variables for reading data and ranges
-  long long minMax[2];
+  long inputData;    // Temporary variables for reading data and ranges
+  long minMax[2];
 
   // Open data file and ranges files at locations given by user
   data = fopen(argv[1],"r");
   ranges = fopen(argv[2],"r");
 
   // Read and insert data values into Binary tree
-  while (fscanf(data,"%lli",&inputData) != EOF) {
+  while (fscanf(data,"%ld",&inputData) != EOF) {
     btreeInsert(tree,inputData);
   }
 
   // Read ranges, query binary tree with range values, and then output
   // if the range query results in an even or odd sum.
-  while (fscanf(ranges,"%lli %lli",&minMax[0],&minMax[1]) != EOF) {
+  while (fscanf(ranges,"%ld %ld",&minMax[0],&minMax[1]) != EOF) {
     if (btreeEvenSumRange(root, minMax[0], minMax[1]) == 1) {
-      printf("Range [%lli,%lli]: %s\n",minMax[0], minMax[1], "odd sum");
+      printf("Range [%ld,%ld]: %s\n",minMax[0], minMax[1], "odd sum");
     }
     else {
-      printf("Range [%lli,%lli]: %s\n",minMax[0], minMax[1], "even sum");
+      printf("Range [%ld,%ld]: %s\n",minMax[0], minMax[1], "even sum");
     }
   }
